Refactor the encrypt_ami module
Move utility functions that are called from multiple places out of
encrypt_ami and into aws_service.  Move constants into a new
aws_constants module.  Remove the old backward compatibility test.  It's
no longer necessary, now that we have a single CLI codebase.
